Engraving: move visibility commands from Score into EditVisibility
Moves cmdSetVisible, cmdUnsetVisible and cmdToggleVisible out of Score
into a new static EditVisibility class. The set-visible/unset-visible
actions now route through a new NotationInteraction::setSelectionVisible
method instead of a Score member-function pointer.
