+++ src/engraving/tests/changevisibility_tests.cpp
@@ -32,6 +32,9 @@
 #include "engraving/dom/hook.h"
 #include "engraving/dom/beam.h"
 
+#include "engraving/editing/editvisibility.h"
+#include "engraving/editing/transaction/transaction.h"
+
 #include "utils/scorerw.h"
 
 #include "log.h"
@@ -128,7 +131,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_SingleNoteChord)
     ASSERT_FALSE(children.empty());
 
     // [WHEN] Hide the note
-    m_score->undoChangeVisible(note, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, false);
 
     // [THEN] Everything in the chord is hidden
     for (EngravingItem* child : children) {
@@ -136,7 +139,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_SingleNoteChord)
     }
 
     // [WHEN] Show the note
-    m_score->undoChangeVisible(note, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, true);
 
     // [THEN] Everything in the chord is visible
     for (EngravingItem* child : children) {
@@ -151,8 +154,8 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_SingleNoteChord)
     ASSERT_TRUE(stem);
 
     // [WHEN] We can also hide the parts of the note
-    m_score->undoChangeVisible(dot, false);
-    m_score->undoChangeVisible(stem, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), dot, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), stem, false);
 
     // [THEN] Everything in the chord is visible, except the parts that were hidden manually
     for (EngravingItem* child : children) {
@@ -176,21 +179,21 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_RestWithDot)
     NoteDot* dot = rest->dotList().front();
 
     // [WHEN] Hide the rest
-    m_score->undoChangeVisible(rest, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), rest, false);
 
     // [THEN] Rest and its dot are hidden
     EXPECT_FALSE(rest->visible());
     EXPECT_FALSE(dot->visible());
 
     // [WHEN] Show the rest
-    m_score->undoChangeVisible(rest, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), rest, true);
 
     // [THEN] Rest and its dot are visible
     EXPECT_TRUE(rest->visible());
     EXPECT_TRUE(dot->visible());
 
     // [WHEN] We can also hide the dot
-    m_score->undoChangeVisible(dot, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), dot, false);
 
     // [THEN] Rest is visible, but the dot is not
     EXPECT_TRUE(rest->visible());
@@ -213,7 +216,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_IgnoredElements)
     ASSERT_TRUE(note);
 
     // [WHEN] Hide the note
-    m_score->undoChangeVisible(note, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, false);
 
     // [THEN] The note is hidden, but the grace notes, lyrics and slur are still visible
     EXPECT_FALSE(note->visible());
@@ -236,7 +239,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_IgnoredElements)
     }
 
     // [WHEN] Show the note
-    m_score->undoChangeVisible(note, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, true);
 
     // [THEN] Everything is visible
     EXPECT_TRUE(note->visible());
@@ -253,7 +256,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_IgnoredElements)
     }
 
     // [WHEN] We can hide any grace note
-    m_score->undoChangeVisible(graceNotes[0], false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), graceNotes[0], false);
 
     // [THEN] Everything is visible except the previously hidden grace note
     EXPECT_TRUE(note->visible());
@@ -291,7 +294,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordContainingSeveral
     ASSERT_TRUE(firstNote);
 
     // [WHEN] Hide the first note
-    m_score->undoChangeVisible(firstNote, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), firstNote, false);
 
     // [THEN] Only the first note is hidden
     EXPECT_FALSE(firstNote->visible());
@@ -316,7 +319,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordContainingSeveral
     ASSERT_TRUE(secondNote);
 
     // [WHEN] Hide the second note
-    m_score->undoChangeVisible(firstNote, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), firstNote, false);
 
     // [THEN] Only the first and the second notes are hidden
     EXPECT_FALSE(firstNote->visible());
@@ -334,7 +337,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordContainingSeveral
 
     // [WHEN] Hide all notes
     for (Note* note : chord->notes()) {
-        m_score->undoChangeVisible(note, false);
+        EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, false);
     }
 
     // [GIVEN] All items attached to this chord
@@ -372,7 +375,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordsConnectedWithBea
 
     // [WHEN] Hide all notes in the first chord
     for (Note* note : firstChord->notes()) {
-        m_score->undoChangeVisible(note, false);
+        EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, false);
     }
 
     // [THEN] Beam/Steam is visible
@@ -405,7 +408,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordsConnectedWithBea
 
         for (EngravingItem* child : children) {
             if (child->isNote()) {
-                m_score->undoChangeVisible(child, false);
+                EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), child, false);
             }
         }
     }
@@ -426,7 +429,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_ChordsConnectedWithBea
     Note* firstChordNote = firstChord->notes().front();
 
     // [WHEN] Show it
-    m_score->undoChangeVisible(firstChordNote, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), firstChordNote, true);
 
     // [THEN] All stems/beam are visible now
     EXPECT_TRUE(beam->visible());
@@ -479,16 +482,16 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_Ornaments)
     ASSERT_TRUE(accidental);
 
     // NOTE invisible makes ORNAMENT invisible
-    m_score->undoChangeVisible(note, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), note, false);
     ASSERT_FALSE(ornament->visible());
 
     // ORNAMENT visible, but note stays invisible
-    m_score->undoChangeVisible(ornament, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), ornament, true);
     ASSERT_TRUE(ornament->visible());
     ASSERT_FALSE(note->visible());
 
     // Ornament accidental invisible, but ornament stays visible
-    m_score->undoChangeVisible(accidental, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), accidental, false);
     ASSERT_FALSE(accidental->visible());
     ASSERT_TRUE(ornament->visible());
 
@@ -504,7 +507,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_Ornaments)
     ASSERT_TRUE(cueNote);
 
     // ORNAMENT invisible, cue note also becomes invisible
-    m_score->undoChangeVisible(ornament, false);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), ornament, false);
     ASSERT_FALSE(ornament->visible());
     ASSERT_FALSE(cueNote->visible());
     for (EngravingItem* el : cueNote->el()) {
@@ -512,7 +515,7 @@ TEST_F(Engraving_ChangeVisibilityTests, UndoChangeVisible_Ornaments)
     }
 
     // CUE NOTE visible, but ornament stays invisible
-    m_score->undoChangeVisible(cueNote, true);
+    EditVisibility::undoChangeVisible(m_score->transactionManager()->currentOrDummyTransaction(), cueNote, true);
     ASSERT_FALSE(ornament->visible());
     ASSERT_TRUE(cueNote->visible());
     for (EngravingItem* el : cueNote->el()) {
@@ -536,7 +539,7 @@ TEST_F(Engraving_ChangeVisibilityTests, CmdToggleVisible)
     m_score->select(measure);
 
     m_score->startCmd(TranslatableString::untranslatable("Change visibility tests"));
-    m_score->cmdToggleVisible();
+    EditVisibility::toggleVisible(m_score->transactionManager()->currentOrDummyTransaction(), m_score);
     m_score->endCmd();
 
     // [THEN] Everything on the first measure is hidden
@@ -558,7 +561,7 @@ TEST_F(Engraving_ChangeVisibilityTests, CmdToggleVisible)
 
     // [WHEN] Call cmdToggleVisible() again
     m_score->startCmd(TranslatableString::untranslatable("Change visibility tests"));
-    m_score->cmdToggleVisible();
+    EditVisibility::toggleVisible(m_score->transactionManager()->currentOrDummyTransaction(), m_score);
     m_score->endCmd();
 
     // [THEN] Everything on the first measure is visible
@@ -576,7 +579,7 @@ TEST_F(Engraving_ChangeVisibilityTests, CmdToggleVisible)
     }
 
     m_score->startCmd(TranslatableString::untranslatable("Change visibility tests"));
-    m_score->cmdToggleVisible();
+    EditVisibility::toggleVisible(m_score->transactionManager()->currentOrDummyTransaction(), m_score);
     m_score->endCmd();
 
     // [THEN] The notes are hidden
@@ -588,7 +591,7 @@ TEST_F(Engraving_ChangeVisibilityTests, CmdToggleVisible)
     m_score->select(measure);
 
     m_score->startCmd(TranslatableString::untranslatable("Change visibility tests"));
-    m_score->cmdToggleVisible();
+    EditVisibility::toggleVisible(m_score->transactionManager()->currentOrDummyTransaction(), m_score);
     m_score->endCmd();
 
     // [THEN] Everything on the first measure is visible
