+++ src/notation/inotationinteraction.h
@@ -53,6 +53,7 @@ public:
 
     // Visibility
     virtual void toggleVisible() = 0;
+    virtual void setSelectionVisible(bool visible) = 0;
 
     // Hit
     virtual EngravingItem* hitElement(const muse::PointF& pos, float width) const = 0;
